# src/main/resources/openapi/schema.yaml
openapi: 3.0.0
info:
  version: '1.0.0'
  title: 'VSA Projekt 2 schema'
  license:
    name: MIT
servers:
  - url: 'localhost:8080/api'
paths:
  /carparks:
    get:
      summary: 'Sample endpoint: Returns carparks'
      operationId: listCarParks
      tags:
        - carPark
      parameters:
        - name: name
          in: query
          description: name of the Car Park
          required: false
          schema:
            type: string
      responses:
        '200':
          description: 'Sample response: Details about a user by ID'
          headers:
            x-next:
              description: A link to the next page of responses
              schema:
                type: string
          content:
            application/json:
              schema:
                type: array
                items:
                    $ref: '#/components/schemas/CarPark'
    post:
        summary: 'Create new car park'
        operationId: createCarPark
        requestBody:
            required: true
            content:
                application/json:
                    schema:
                        $ref: '#/components/schemas/CarPark'
        responses:
            '201':
                description: 'Entity created and returned'
                content:
                    application/json:
                        schema:
                            type: array
                            items:
                                type: object
                                $ref: '#/components/schemas/CarPark'
            '400':
                description: 'Incorrectly formed body'
  /carparks/{id}:
    get:
        summary: Get carp park by id
        operationId: getCarParkById
        parameters:
            - name: id
              in: path
              description: 'id of the car park'
              required: true
              schema:
                  type: integer
                  format: int64
        responses:
            '200': 
                description: 'Car park was found and returned'
                content:
                    application/json:
                        schema: 
                            $ref: '#/components/schemas/CarPark'
            '404':
                description: 'Car park with id not found'
    put:
        summary: 'Update car park'
        operationId: updateCarPark
        parameters:
            - name: id
              in: path
              description: 'id of the car park'
              required: true
              schema:
                  type: integer
                  format: int64
        requestBody:
            required: true
            content:
                application/json:
                    schema:
                        $ref: '#/components/schemas/CarPark'
        responses:
            '200':
                description: 'Car park was updated and returned'
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/CarPark'
            '400':
                description: 'Ijncorrectly formed body'
            '404':
                description: 'Car park with id was not found'

    delete:
        summary: 'Delete Car park along with all floors and spots'
        operationId: deleteCarPark
        parameters:
            - name: id
              in: path
              required: true
              schema:
                  type: integer
                  format: int64
        responses:
            '204':
                description: 'Car park was deleted'
            '404':
                description: 'Car park with id was not found'
  /carparks/{id}/floors:
    get:
        summary: 'Return collection of all floors for car park with given id'
        operationId: getCarParkFloors
        parameters:
            - name: id
              in: path
              required: true
              schema:
                  type: integer
                  format: int64
        responses:
            '200':
                description: 'Returns all car park floors for given car park'
                content:
                    application/json:
                        schema:
                            type: array
                            items:
                                $ref: '#/components/schemas/CarParkFloor'
            '404':
                description: 'Car park with given id does not exist so there are no Car Park floors for it'
    post:
        summary: 'Create new car park floor for given car park'
        operationId: createCarParkFloor
        parameters:
            - name: id
              required: true
              in: path
              description: 'Id of car park'
              schema:
                  type: integer
                  format: int64
        requestBody:
            required: true
            content:
                application/json:
                    schema:
                        $ref: '#/components/schemas/CarParkFloor'
        responses:
            '201':
                description: 'Car park floor created and returned'
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/CarParkFloor'
            '404':
                description: 'Car park does not exist and car park floor cannot be created for it'
            '400':
                description: 'Incorrectly formed request body'
...
  /carparks/{id}/floors/{identifier}/spots:
    get:
        summary: 'Get parking spots for given car park and given floor'
        operationId: getParkingSpotsForFloor
        parameters:
            - name: id
              in: path
              required: true
              schema:
                  type: integer
                  format: int64
              description: 'Id of car park'
            - name: identifier
              in: path
              required: true
              description: 'Identifier of the car park floor'
              schema:
                  type: string
        responses:
            '200':
                description: 'Car park and floor exist and parking spots are returned'
                content:
                    application/json:
                        schema:
                            type: array
                            items:
                                $ref: '#/components/schemas/ParkingSlot'
            '404':
                description: 'Car park or car park floor does not exist'
    
    post:
        summary: 'Create parking spot for given car park and given car park floor'
        operationId: createParkingSpot
        parameters:
            - name: id
              in: path
              required: true
              schema:
                  type: integer
                  format: int64
              description: 'Id of car park'
            - name: identifier
              in: path
              required: true
              description: 'Identifier of the car park floor'
              schema:
                  type: string
        requestBody:
            required: true
            content:
                application/json:
                    schema:
                        $ref: '#/components/schemas/ParkingSpot'
        responses:
            '201':
                description: 'Parking spot created and returned'
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/ParkingSpot'
            '404':
                description: 'Car park or car park floor does not exist and parking floor cannot be created for it'
            '400':
                description: 'Incorrectly formed request body'
...
  /cars/{id}:
        get:
            summary: 'Get car with id'
            operationId: getCar
            parameters:
                - name: id
                  in: path
                  required: true
                  description: 'Id of car'
                  schema:
                      type: integer
                      format: int64
            responses:
                '200':
                    description: 'Car exists and is returned'
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Car'
                '404':
                    description: 'Car with given id does not exist'
        put:
            summary: 'Update car'
            operationId: updateCar
            parameters:
                - name: id
                  in: path
                  required: true
                  description: 'Id of car'
                  schema:
                      type: integer
                      format: int64
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/Car'
            responses:
                '201':
                    description: 'Car was created and returned'
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Car'
                '404':
                    description: 'Car was not found'
                '400':
                    description: 'Incorrectly formed request body'
        delete:
            summary: 'Delete car'
            operationId: deleteCar
            parameters:
                - name: id
                  in: path
                  required: true
                  description: 'Id of car'
                  schema:
                      type: integer
                      format: int64
            responses:
                '204':
                    description: 'Car deleted'
                '404':
                    description: 'Car does not exist'
...
  /cartypes:
        get:
            summary: 'Get car types'
            operationId: getCarTypes
            parameters:
                - name: name
                  required: false
                  in: query
                  description: 'Name of a cartype. If set only type with that name will be part of the collection'
                  schema:
                      type: string
            responses:
                '200':
                    description: 'Car types returned'
                    content:
                        application/json:
                            schema:
                                type: array
                                items:
                                    $ref: '#/components/schemas/CarType'
        
        post:
            summary: 'Create car type'
            operationId: createCarType
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                           $ref: '#/components/schemas/CarType'
            responses:
                '201':
                    description: 'Car type created and returned'
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/CarType'
                '400':
                    description: 'Incorrectly formed request body'
  /cartypes/{id}:
        get:
            summary: 'Get car type'
            operationId: getCarType
            parameters:
                - name: id
                  required: true
                  in: path
                  description: 'Id of the car type'
                  schema:
                      type: integer
                      format: int64
            responses:
                '200':
                    description: 'Car type found and returned'
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/CarType'
                '404':
                    description: 'Car type does not exist'
        delete:
            summary: 'Delete car type'
            operationId: deleteCarType
            parameters:
                - name: id
                  required: true
                  in: path
                  description: 'Id of the car type'
                  schema:
                      type: integer
                      format: int64
            responses:
                '204':
                    description: 'Car type deleted'
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/CarType'
                '404':
                    description: 'Car type not found'
...
components:
  schemas:
    CarPark:
        type: object
        required:
            - id
            - name
            - address
            - prices
        properties:
            id:
                type: integer
                format: int64
            name:
                type: string
            address:
                type: string
            prices:
                type: number
            floors:
                type: array
                items:
                    $ref: CarParkFloor
    CarParkFloor:
        type: object
        required:
            - id
            - identifier
            - carPark
        properties:
            id:
                type: integer
                format: int64
            identifier:
                type: string
            carPark:
                type: integer
                format: int64
            spots:
                type: array
                items:
                    typ: object
                    $ref: ParkingSpot
    ParkingSpot:
        type: object
        required:
            - id
            - identifier
            - carParkFloor
            - carPark
            - free
        properties:
            id:
                type: integer
                format: int64
            identifier:
                type: string
            carParkFloor:
                type: string
            carPark:
                type: integer
                format: int64
            type:
                type: object
                $ref: CarType  
            free:
                type: boolean
            reservations:
                type: array
                items:
                    $ref: ReservationDownFromParkingSpot
...
    Car:
        type: object
        required:
            - id
            - brand
            - model
            - vrp
            - color
            - owner
        properties:
            id:
                type: integer
                format: int64
            brand:
                type: string
            model:
                type: string
            vrp:
                type: string
            colour:
                type: string
            type:
                type: object
                $ref: CarType
            owner:
                type: object
                $ref: UserDownFromCar
            reservations:
                type: array
                items:
                    type: object
                    $ref: ReservationDownFromCar
...
    CarType:
        type: object
        required:
            - id
            - name
        properties:
            id:
                type: integer
                format: int64
            name:
                type: string
...
  securitySchemes:
    BasicAuth:
      type: http
      scheme: basic
security:
  - BasicAuth: []